+++ NEWS.md
@@ -1,3 +1,6 @@
+2020-01-07
+* GW-278: Added a new class that runs a timer and checks twice a day (every 12 hours) whether or not there are products in the products folder that are 10 days or older. If so, they are removed from the respective products folder.
+
 2020-01-03
 * (FIX) GW-205: FC is validated without affecting other obscuration phenomena
 - used the string "funnel clouds" everywhere in the code.